requirements.txt
chromadb==0.4.18
# pysqlite3==0.5.2
python-dotenv==1.0.0
Requests==2.31.0
services==0.1.1
streamlit==1.26.0
pysqlite-binary
openai
Flask